requirements.txt
streamlit>=1.56.0
reportlab>=4.4.0
pillow>=10.0.0
anthropic>=0.94.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
